added logic to render scientific muscle name where colloquial name unavailable. Pulled out reusable muscle group filter function into its own component
